requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
web3>=6.15.0
reportlab>=4.0.9
pycryptodome>=3.20.0
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.3
qrcode>=7.4.2
pillow>=10.4.0
PyPDF2>=3.0.1
pydantic>=2.5.3
aiofiles>=23.2.1
cryptography>=42.0.0
